requirements.txt
celery-connectors
colorlog
dj-database-url==0.4.2
Django==2.0
django-admin
django-configurations==2.0.0
django-debug-toolbar==1.9.1
django-extensions==1.9.7
django-rest-registration
django-rest-swagger
django-six
djangorestframework
djangorestframework-jwt
flake8
gunicorn==19.7.1
jsonfield
keras
network-pipeline
newrelic
pandas
numpy
psycopg2==2.7.3.2
pycodestyle
pydocstyle
requests
scikit-learn
tensorflow
virtualenvwrapper
whitenoise==4.0b4
Werkzeug==0.12.2
